% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_snippets.R
\name{add_snippets}
\alias{add_snippets}
\title{add Shiny argonDash snippets to Rstudio snippets file}
\usage{
add_snippets(rs = "~/.R/snippets/r.snippets")
}
\arguments{
\item{rs}{path to rstudio snippets file default: "~/.R/snippets/r.snippets"}
}
\description{
add Shiny argonDash snippets to Rstudio snippets file
}
\examples{
argonSnippets::add_snippets()
}
